Ajustes finais

+++ src/pages/Remover.js
@@ -5,16 +5,27 @@ import Carregando from '../components/Carregando';
 import Mensagem from '../components/Mensagem';
 
 const ResultadoRemover = () => {
+  const [erro, setErro] = useState('');
   const { tipo, id } = useRouteMatch().params;
   const [sucesso, setSucesso] = useState(null);
 
   useEffect(() => {
     if(tipo === 'aluno' || tipo === 'professor') {
-      firebaseDB.firestore().collection('usuarios').doc(id).delete().then(function() {
-        setSucesso(true);
-      }).catch(function(error) {
+      firebaseDB.firestore().collection('membros').where('usuario', '==', id).get().then((snapshot) => {
+        if(snapshot.size >= 1) {
+          setErro('Você precisa primeiro desvincular o usuário de projetos.');
+          throw new Error('Você precisa primeiro desvincular o usuário de projetos.');
+        } else {
+          firebaseDB.firestore().collection('usuarios').doc(id).delete().then(function() {
+            setSucesso(true);
+          }).catch(function(error) {
+            setSucesso(false);
+            console.error("Erro ao remover cadastro!", error);
+          });
+        }
+      }).catch(function (error) {
         setSucesso(false);
-        console.error("Erro ao remover cadastro!", error);
+        console.error("Erro ao carregar membros!", error);
       });
     } else if(tipo === 'projeto') {
       firebaseDB.firestore().collection('projetos').doc(id).delete().then(function() {
@@ -36,7 +47,7 @@ const ResultadoRemover = () => {
   if(sucesso === true) {
     return <Mensagem titulo="Removido com sucesso!" texto="Entrada removida com seucesso do banco de dados." />;
   } else if (sucesso === false) {
-    return <Mensagem titulo="Erro ao remover!" texto="Houve um erro ao tentar remover a entrada do banco de dados." />;
+    return <Mensagem titulo="Erro ao remover!" texto={ erro === '' ? "Houve um erro ao tentar remover a entrada do banco de dados." : erro } />;
   } else {
     return <Carregando />;
   }
